Encrypt keys before passing through to backend store.
We don't trust the backend database to securely store keys on behalf of
clients. The C2 acts as an encrypting layer through to the database. We
add passphrase, a setting in the configuration file whose value is
passed through argon2 as a KDF to produce a KEK to encrypt keys.
Keys are then encrypted with existing AES-SIV code.