Introduce tests to exercise DB.go using sqlite3.
Testing code for the database is introduced by this commit, which
creates many-to-many links in a number of scenarios, adds, deletes and
tests the results with raw sql.

Files are created in /tmp/e4c2_unittest_randomjunk.sqlite and are
cleaned up if the test succeeds.

A script is provided to clean the go test cache and run unit tests on
all projects that are present.