Dockerize the C2 backend. This commit fixes issues in the build of the docker container. If BUILDVERSION is not supplied we tag the image with "devel" instead, along with tagging for a given git commit.
The backend now looks for config files in ../configs relative to the
directory it is launched from first. This allows deployment in
$E4DIR with the usual unix layout.

The build script is fixed to warn immediately if git is not present, the
git repository does not exist (happens in our docker build environment).
It will also check for E4_GIT_VERSION and rely on this if present, to
allow the docker build environment to specify it.